Add front hallway light automation
The hallway now detects occupancy via a single motion sensor
and turns the light on or off accordingly. Only at night, because even
though the motion sensor says it is dark it is actually bright enough
during the day and we don't need the additional light just for
passing through.