Add presence detection to the living room
A simple bayesian sensor seems to be good enough to differentiate
between somebody actually being in the living room versus the
adjacent dining room or the cat having a party at night.

I have the feeling the number of scenes is slowly getting out of
control and it will only get worse for every room that gets some
lighting automations so I should think about how I could simplify
this a bit.