Turn on dimmed light when the TV is on
When the TV is on and the living room is occupied the living room
changes to On and Dimmed. This prevents the light from turning on
and off all the time when we are watching TV with the light off when it
gets dark and the motion sensors register something from time to time.