Living room goes to Occupied when using the switch
When the TV in the living room is on and the button on the light switch
is pressed while the room is occupied it now goes to state Occupied
instead of Off. This way the light stays on but flips to off as soon as the
room is not occupied anymore.

Also reduce the brightnes of the lights when the living room is occupied
to clearly show that the button has been pressed. This has the strange
effect that the light is actually getting brighter when they are turned
off, when the living room changes from On and Dimmed to
Occupied. Maybe I will find a solution for that someday.

Implements #26 for the living room.