Make music in the kitchen play automatically
I'm still not sure if we *really* want this but let's try for a couple of days. When there's music playing in the dining room and the kitchens flips to Occupied or On the music is also played there.

The other way around is currently not viable because there are no realiable occupancy sensors in the dining room.

Closes #29