Soft transfer for lights under the bed
When the mood light was turned off and later the night lights came on the lights were the last color of the mood lights for like half a second befor switching to the night light which was not nice in the middle of the night.

The workaround is to always change to the night light before turning the mood light off.