Completely overhault living room states
I didn't test this at all so we're living in a haunted house until I fixed everything that's wrong.

For an overview of the new state changes see this handy flowchart on my (completely german, sorry!) blog: https://blog.frdnspnzr.de/2020/04/was-hat-zwei-daumen-und-einen-irrational-komplexen-lichtschalter-gebaut/

This eliminates the living room modes and closes #30 as 'Dimmed' is now a state.