Add experimental freeze warning
I want HA to alert us when it will freeze tonight so we can get some flowers inside. The temperatures we got until now were several degrees off from the YR app on my phone so the information was more or less worthless. I don't want to depend on this alert until I know for sure that the temperatures are correct so now I added a bit of debugging information.