Make living room light dependent on light levels
Sun state is a *mostly* fine indicator but weird things kept happening in edge cases. The light level of the living room is what I actually want to know. Because the sensors are in various parts of the room the maximum light level is used. Some adjustements for the best light level are probably needed.

There's still the weird edge case of light levels of exactly 10 lux because both below and above are inclusive. I have no idea how to adress this acceptable effort and I have the strong oppinion that this should be changed in HA.