Add reshuffling to Sonos speakers
I have no idea why, but somebody at Sonos thought it would be a great idea to just reset play position after some time. Now when we start a long playlist in shuffle, stop the music and restart after some time it just plays the same songs again because it starts with the first song in the same order.

These scripts circumvent this behaviour be reshuffling the saved playlist of the Sonos speakers after ten minutes of being paused.