Add speech notifications to state changes
My wife was complaining that she didn't always know wether she *really* turned the lights off or not when the room is still illuminated because it is occupied (and she is absolutely right about that to be honest). So why not just tell her?

Sadly I can't just react to state changes because I don't want the speech script to execute every time the room goes from occupied to off or the like but only on button presses.