Simplify TV backlighting automations
Most states for the backlighting are not dependent on the TV. When the TV is on WLED gets the signal to accept remote signals from Hyperion and Hyperion just overwrites whatever WLED is supposed to do. In the background the state of WLED still changes but that has no effect. When the TV is turned off again WLED stops accepting remote signals and returns to whatever is configured at that moment.