Add basic darkness indicators
This is in preparation of better room light handling. Instead of relying on sunrise and sundown to know if I should turn on the lights or not I'm going to add some more complex logic in the future.

For now I will only add the boolean inputs and see wether the results are sensible or not. Later I will use these inputs instead of sun state in automations.