Add Teams status to HA
My setup closely resembles this guide: https://community.home-assistant.io/t/microsoft-teams-status/202388/38

A Powershell script running on my computer watches the Microsoft Teams log file for changes in status and reports that to HA as sensor.teams_raw. The template sensor sensor.teams parses that into a more useful format. The Powershell script is not included here.

Right now the lighting in my study changes according to my status as a proof of concept. I'm going to implement automations so a light in the living room turns red when I'm in call so my wife knows when she can come in.