Add automations for bedroom TV
We got an old TV for our bedroom and put a Chromecast in to make it smart. The remote to the TV is missing so we currently have no way to turn it on or off from the bed.

The more relevant of the two is turning the TV off as we can just walk up to the TV and turn it on before going to bed. So I put a smart plug there and an IKEA Shortcut Button on our nightstand so we can turn the TV off from our bed. Problem solved.